+++ include/zenoh-pico/transport/unicast/lease.h
@@ -18,8 +18,8 @@
 #include "zenoh-pico/transport/transport.h"
 
 int8_t _zp_unicast_send_keep_alive(_z_transport_unicast_t *ztu);
-int _zp_unicast_start_lease_task(_z_transport_t *zt, _z_task_attr_t *attr, _z_task_t *task);
-int _zp_unicast_stop_lease_task(_z_transport_t *zt);
+int8_t _zp_unicast_start_lease_task(_z_transport_t *zt, _z_task_attr_t *attr, _z_task_t *task);
+int8_t _zp_unicast_stop_lease_task(_z_transport_t *zt);
 void *_zp_unicast_lease_task(void *ztu_arg);  // The argument is void* to avoid incompatible pointer types in tasks
 
 #endif /* ZENOH_PICO_TRANSPORT_LINK_TASK_LEASE_H */
