+++ src/transport/multicast/read.c
@@ -39,7 +39,7 @@ int8_t _zp_multicast_read(_z_transport_multicast_t *ztm) {
     return ret;
 }
 
-int _zp_multicast_start_read_task(_z_transport_t *zt, _z_task_attr_t *attr, _z_task_t *task) {
+int8_t _zp_multicast_start_read_task(_z_transport_t *zt, _z_task_attr_t *attr, _z_task_t *task) {
     // Init memory
     (void)memset(task, 0, sizeof(_z_task_t));
     // Attach task
@@ -53,7 +53,7 @@ int _zp_multicast_start_read_task(_z_transport_t *zt, _z_task_attr_t *attr, _z_t
     return _Z_RES_OK;
 }
 
-int _zp_multicast_stop_read_task(_z_transport_t *zt) {
+int8_t _zp_multicast_stop_read_task(_z_transport_t *zt) {
     zt->_transport._multicast._read_task_running = false;
     return _Z_RES_OK;
 }
@@ -143,14 +143,14 @@ int8_t _zp_multicast_read(_z_transport_multicast_t *ztm) {
     return _Z_ERR_TRANSPORT_NOT_AVAILABLE;
 }
 
-int _zp_multicast_start_read_task(_z_transport_t *zt, _z_task_attr_t *attr, _z_task_t *task) {
+int8_t _zp_multicast_start_read_task(_z_transport_t *zt, _z_task_attr_t *attr, _z_task_t *task) {
     _ZP_UNUSED(zt);
     _ZP_UNUSED(attr);
     _ZP_UNUSED(task);
     return _Z_ERR_TRANSPORT_NOT_AVAILABLE;
 }
 
-int _zp_multicast_stop_read_task(_z_transport_t *zt) {
+int8_t _zp_multicast_stop_read_task(_z_transport_t *zt) {
     _ZP_UNUSED(zt);
     return _Z_ERR_TRANSPORT_NOT_AVAILABLE;
 }
